feat: BREAKING CHANGE
Added support for proxmox_virtual_environment_download_file as the container template file resource. If using a URL/remote image as the template, it will delete the original template file that was created and re-download the image as a new container template. This will also cause and containers created using the original proxmox_virtual_environment_file resource to be deleted and re-created.

To bypass the backing template file replacement, set the module input variable "migrate_template_file" to false. This is by default set to true so any future resources created will use the newer resource, and any resources this is not set on to delete and recreate the backing template and the containers referencing it.